+++ src/main/kotlin/click/seichi/gigantic/listener/PlayerMonitor.kt
@@ -119,58 +119,10 @@ class PlayerMonitor : Listener {
                 GiganticEventMessages.DROPPED_RELIC(Relic.CUP_OF_KING).sendTo(player)
             }
         }
-        //todo:生成方法を細かくするべき
-        //ミッション用コード
-        if (player.getOrPut(Keys.MISSION_MAP).values.any { it.date != DateTime.now().withTimeAtStartOfDay() }) {
-            player.transform(Keys.MISSION_MAP) {
-                it.toMutableMap().apply {
-                    entries.removeIf { entry -> entry.value.date != DateTime.now().withTimeAtStartOfDay() }
-                }
-            }
-        }
-        val missionCount = player.getOrPut(Keys.MISSION_MAP).values.map { it.missionType == 1 }
-        if (missionCount.size < Config.MISSION_DAILY_AMOUNT) {
-            MissionMessages.MISSION_CREATE.sendTo(player)
-            missionCreate(player,1)
-        }
+        //ミッション生成コード
+        Mission.missionCreate(player, 1)
     }
 
-    fun missionCreate(player: Player,missionType: Int){
-        var index = 0
-        val missionIds = mutableSetOf<Int>()
-        val digest = MessageDigest.getInstance("SHA-256")
-        while (missionIds.size < Config.MISSION_DAILY_AMOUNT) {
-        val concatenatedString = "${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_$index"
-        val random = Random(concatenatedString.hashCode().toLong())
-        missionIds.add(random.nextInt(1, Mission.values().size + 1))
-        index++
-    }
-        for (i in 0 until Config.MISSION_DAILY_AMOUNT) {
-            val newMission = MissionClient(
-                missionId = missionIds.elementAt(i),
-                missionType = missionType,
-                missionDifficulty = Random("${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_${i}_missionDifficulty".hashCode().toLong()).nextInt(0, 3),
-                missionReqSize = if (missionIds.elementAt(i) == 6){
-                    Random("${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_${i}_missionDifficulty".hashCode().toLong()).nextInt(0, Mission.RequestWillSize.values().size)
-                } else 0,
-                missionReqBlock = if (missionIds.elementAt(i) == 3) {
-                    Random("${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_missionReqBlock".hashCode().toLong()).nextInt(0, Mission.RequestBlockType.values().size)
-                } else 0,
-                progress = 0.0,
-                complete = false,
-                rewardReceived = false,
-                date = DateTime.now()
-            )
-            for (j in 0 until 12){
-                info("${Random("${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_${j}_missionDifficulty".hashCode().toLong()).nextInt(1, 4)}")
-            }
-            player.transform(Keys.MISSION_MAP) {
-                it.toMutableMap().apply {
-                    put(newMission.missionId, newMission)
-                }
-            }
-        }
-    }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     fun onBlockBreak(event: BlockBreakEvent) {
