+++ src/main/kotlin/click/seichi/gigantic/mission/Mission.kt
@@ -3,6 +3,7 @@ package click.seichi.gigantic.mission
 import click.seichi.gigantic.message.LocalizedText
 import click.seichi.gigantic.will.WillSize
 import click.seichi.gigantic.cache.key.Keys
+import click.seichi.gigantic.config.Config
 import click.seichi.gigantic.extension.*
 import click.seichi.gigantic.message.messages.MissionMessages
 import click.seichi.gigantic.player.ToggleSetting
@@ -10,7 +11,9 @@ import click.seichi.gigantic.sound.sounds.PlayerSounds
 import org.bukkit.Material
 import org.bukkit.block.Block
 import org.bukkit.entity.Player
+import org.joda.time.DateTime
 import java.util.*
+import kotlin.random.Random
 
 /**
  * @author 2288-256
@@ -108,6 +111,49 @@ enum class Mission(
                         MissionProgressManager.spawn(mission,missionProgressBar)
                     }
                 }
+            }
+        }
+        fun missionCreate(player: Player,missionType: Int) {
+            if (player.getOrPut(Keys.MISSION_MAP).values.any { it.date != DateTime.now().withTimeAtStartOfDay() }) {
+                player.transform(Keys.MISSION_MAP) {
+                    it.toMutableMap().apply {
+                        entries.removeIf { entry -> entry.value.date != DateTime.now().withTimeAtStartOfDay() }
+                    }
+                }
+            }
+            val missionCount = player.getOrPut(Keys.MISSION_MAP).values.map { it.missionType == 1 }
+            if (missionCount.size < Config.MISSION_DAILY_AMOUNT) {
+                var index = 0
+                val missionIds = mutableSetOf<Int>()
+                while (missionIds.size < Config.MISSION_DAILY_AMOUNT) {
+                    val concatenatedString = "${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_$index"
+                    val random = Random(concatenatedString.hashCode().toLong())
+                    missionIds.add(random.nextInt(1, Mission.values().size + 1))
+                    index++
+                }
+                for (i in 0 until Config.MISSION_DAILY_AMOUNT) {
+                    val newMission = MissionClient(
+                        missionId = missionIds.elementAt(i),
+                        missionType = missionType,
+                        missionDifficulty = Random("${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_${i}_missionDifficulty".hashCode().toLong()).nextInt(0, 3),
+                        missionReqSize = if (missionIds.elementAt(i) == 6){
+                            Random("${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_${i}_missionDifficulty".hashCode().toLong()).nextInt(0, Mission.RequestWillSize.values().size)
+                        } else 0,
+                        missionReqBlock = if (missionIds.elementAt(i) == 3) {
+                            Random("${player.uniqueId}_${DateTime.now().withTimeAtStartOfDay()}_missionReqBlock".hashCode().toLong()).nextInt(0, Mission.RequestBlockType.values().size)
+                        } else 0,
+                        progress = 0.0,
+                        complete = false,
+                        rewardReceived = false,
+                        date = DateTime.now()
+                    )
+                    player.transform(Keys.MISSION_MAP) {
+                        it.toMutableMap().apply {
+                            put(newMission.missionId, newMission)
+                        }
+                    }
+                }
+                MissionMessages.MISSION_CREATE.sendTo(player)
             }
         }
     }
